Remove dead analyzer
