--- src/Immediate.Jobs.Analyzers/DiagnosticDescriptors.cs
@@ -1,48 +1,0 @@
-using Microsoft.CodeAnalysis;
-
-namespace Immediate.Jobs.Analyzers;
-
-/// <summary>Diagnostic descriptors produced for invalid job declarations.</summary>
-internal static class DiagnosticDescriptors
-{
-	private const string Category = "Immediate.Jobs";
-
-	public static readonly DiagnosticDescriptor UnsupportedPayload = Create(
-		"IJOB003",
-		"Unsupported job payload",
-		"Payload type '{0}' cannot be source-generated for JSON serialization: {1}"
-	);
-
-	public static readonly DiagnosticDescriptor InvalidMethodSignature = Create(
-		"IJOB004",
-		"Invalid job method signature",
-		"Job '{0}' must declare exactly one private instance HandleAsync method returning ValueTask, with a request followed by CancellationToken"
-	);
-
-	public static readonly DiagnosticDescriptor InvalidQueueConfiguration = Create(
-		"IJOB010",
-		"Invalid queue configuration",
-		"Queue '{0}' has invalid configuration: {1}"
-	);
-
-	public static readonly DiagnosticDescriptor InvalidQueueTarget = Create(
-		"IJOB011",
-		"Invalid queue target",
-		"Queue type '{0}' must be marked with QueueDefinitionAttribute"
-	);
-
-	public static readonly DiagnosticDescriptor InvalidContextExtractor = Create(
-		"IJOB013",
-		"Invalid job context extractor",
-		"Context extractor type '{0}' must implement exactly one IJobContextExtractor<TContext> interface"
-	);
-
-	public static readonly DiagnosticDescriptor UnsupportedContext = Create(
-		"IJOB014",
-		"Unsupported job context",
-		"Context type '{0}' cannot be source-generated for JSON serialization: {1}"
-	);
-
-	private static DiagnosticDescriptor Create(string id, string title, string message) =>
-		new(id, title, message, Category, DiagnosticSeverity.Error, isEnabledByDefault: true);
-}
