--- src/Immediate.Jobs.Analyzers/ImmediateJobsAnalyzer.cs
@@ -1,140 +1,0 @@
-using System.Collections.Immutable;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
-
-namespace Immediate.Jobs.Analyzers;
-
-/// <summary>Validates declarations marked with <c>Immediate.Jobs.Shared.JobAttribute</c>.</summary>
-[DiagnosticAnalyzer(LanguageNames.CSharp)]
-public sealed class ImmediateJobsAnalyzer : DiagnosticAnalyzer
-{
-	/// <inheritdoc />
-	public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } =
-		ImmutableArray.Create(
-			DiagnosticDescriptors.UnsupportedPayload,
-			DiagnosticDescriptors.InvalidMethodSignature,
-			DiagnosticDescriptors.InvalidQueueConfiguration,
-			DiagnosticDescriptors.InvalidQueueTarget,
-			DiagnosticDescriptors.InvalidContextExtractor,
-			DiagnosticDescriptors.UnsupportedContext
-		);
-
-	/// <inheritdoc />
-	public override void Initialize(AnalysisContext context)
-	{
-		ArgumentNullException.ThrowIfNull(context);
-		context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
-		context.EnableConcurrentExecution();
-		context.RegisterCompilationAction(AnalyzeCompilation);
-	}
-
-	private static void AnalyzeCompilation(CompilationAnalysisContext context)
-	{
-		AnalyzeQueues(context);
-		var jobs = JobDiscovery.FindJobs(context.Compilation, context.CancellationToken);
-
-		var hasNodaTimeIntegration = context.Compilation.ReferencedAssemblyNames
-			.Any(identity => identity.Name == "Immediate.Jobs.NodaTime");
-		foreach (var job in jobs)
-		{
-			context.CancellationToken.ThrowIfCancellationRequested();
-			var location = GetAttributeLocation(job);
-			//var contextUses = JobDiscovery.GetJobContextUses(job);
-			//foreach (var contextUse in contextUses)
-			//{
-			//	var contextLocation = contextUse.AppliedAttribute.ApplicationSyntaxReference?.GetSyntax().GetLocation() ?? location;
-			//	if (contextUse.ContextType is null)
-			//	{
-			//		context.ReportDiagnostic(Diagnostic.Create(
-			//			DiagnosticDescriptors.InvalidContextExtractor,
-			//			contextLocation,
-			//			contextUse.ExtractorType.ToDisplayString()
-			//		));
-			//		continue;
-			//	}
-
-			//	var contextProblem = PayloadValidation.FindProblem(contextUse.ContextType);
-			//	if (contextProblem is not null)
-			//	{
-			//		context.ReportDiagnostic(Diagnostic.Create(
-			//			DiagnosticDescriptors.UnsupportedContext,
-			//			contextLocation,
-			//			contextUse.ContextType.ToDisplayString(),
-			//			contextProblem
-			//		));
-			//	}
-			//}
-
-			if (JobDiscovery.GetUsesQueueAttribute(job) is { AttributeClass.TypeArguments: [INamedTypeSymbol queueType] } usesQueue &&
-				JobDiscovery.GetQueueDefinitionAttribute(queueType) is null)
-			{
-				context.ReportDiagnostic(Diagnostic.Create(
-					DiagnosticDescriptors.InvalidQueueTarget,
-					usesQueue.ApplicationSyntaxReference?.GetSyntax().GetLocation() ?? location,
-					queueType.ToDisplayString()
-				));
-				continue;
-			}
-
-			var methods = job.GetMembers("HandleAsync").OfType<IMethodSymbol>()
-				.Where(method => !method.IsImplicitlyDeclared)
-				.ToArray();
-			if (methods.Length != 1 || !JobDiscovery.IsValidMethod(methods[0], out var payloadType, out var hasPayload))
-			{
-				context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidMethodSignature, location, job.Name));
-				continue;
-			}
-
-			// cron, time zone, and the remaining `[Job]` values are validated by JobClassAnalyzer
-
-			if (!hasPayload)
-				continue;
-
-			var problem = PayloadValidation.FindProblem(payloadType!);
-			if (problem is not null)
-			{
-				context.ReportDiagnostic(Diagnostic.Create(
-					DiagnosticDescriptors.UnsupportedPayload,
-					methods[0].Parameters[0].Locations.FirstOrDefault() ?? location,
-					payloadType!.ToDisplayString(),
-					problem
-				));
-			}
-		}
-	}
-
-	private static void AnalyzeQueues(CompilationAnalysisContext context)
-	{
-		var queues = JobDiscovery.FindQueueDefinitions(context.Compilation, context.CancellationToken);
-		foreach (var queue in queues)
-		{
-			var attribute = JobDiscovery.GetQueueDefinitionAttribute(queue)!;
-			var name = JobDiscovery.GetQueueName(queue);
-			string? problem = null;
-			if (string.IsNullOrWhiteSpace(name))
-				problem = "Name cannot be empty";
-			else if (name == "default")
-				problem = "Name 'default' is reserved";
-			else if (JobDiscovery.GetNamedInt(attribute, "Concurrency", 0) < 0)
-				problem = "Concurrency cannot be negative";
-
-			if (problem is not null)
-			{
-				context.ReportDiagnostic(Diagnostic.Create(
-					DiagnosticDescriptors.InvalidQueueConfiguration,
-					attribute.ApplicationSyntaxReference?.GetSyntax().GetLocation() ?? queue.Locations.FirstOrDefault(),
-					queue.Name,
-					problem
-				));
-			}
-		}
-	}
-
-	private static Location GetAttributeLocation(INamedTypeSymbol symbol)
-	{
-		var attribute = JobDiscovery.GetJobAttribute(symbol);
-		return attribute?.ApplicationSyntaxReference?.GetSyntax().GetLocation()
-			?? symbol.Locations.FirstOrDefault()
-			?? Location.None;
-	}
-}
